Fix 'cu' VIM mapping
In case the mapping is executed on the first word of a line that contains more words, the pasting at the end would have added an extra line.
Fix this by pasting with 'P' instead of 'p'. This required using <c-o> instead of <esc>

Also make 'cu' keep the cursor at the same position, where previously it was moved to the end of the word